# application/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Programmer Notebook
  version: 1.0.0
servers:
  - url: https://0.0.0.0:8080/api/v2
...
paths:
...
  /users:
    get:
      tags:
        - users
      summary: Get all users in database
      description: Получение всех зарегестрированных пользователей
      operationId: getAllUsers

      security:
        - JWTAuth: []

      responses:
        "200":
          description: Success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPublicInfo'
        "403":
          description: Недостаточно прав
        "500":
          description: Internal error
  /users/{id}:
    get:
      tags:
        - users
      summary: Get user
      description: Получение одного пользователя
      operationId: getUser


      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserFullInfo'
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such user
        "500":
          description: Internal error

    delete:
      tags:
        - users
      summary: Delete user
      description: Удаление одного пользователя
      operationId: deleteUser

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Success
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: Пользователь не найден
        "500":
          description: Внутренняя ошибка сервера

    patch:
      tags:
        - users
      summary: Update user
      description: Изменение данных пользователя
      operationId: updateUser

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: "ID, или логин, или ФИО изменяемого пользователя"
          required: true
          schema:
            type: integer

      requestBody:
        description: Новые данные пользователя
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UserPrivateInfo'
        required: true

      responses:
        "200":
          description: Success
          content: {}
        "401":
          description: Authorization error
          content: {}
        "403":
          description: Недостаточно прав
          content: {}
        "404":
          description: No such user
          content: {}
        "500":
          description: Internal error
          content: {}

      #####################################################################
      # Team API
      #####################################################################
...
components:
  schemas:
...
    UserPublicInfo:
      required:
        - id
        - fio
        - role
      type: object
      properties:
        id:
          type: integer
          format: int
          example: 10
        fio:
          type: string
          example: Иванов Иван Иванович
        role:
          type: integer
          description: "Роль пользователя (Админ, Читатель, Автор)"
          format: int
          example: 1
    UserPrivateInfo:
      required:
        - id
        - fio
        - login
        - password
        - role
      type: object
      properties:
        id:
          type: integer
          format: int
          example: 10
        fio:
          type: string
          example: Иванов Иван Иванович
        login:
          type: string
          example: login
        password:
          type: string
          example: password
        role:
          type: integer
          description: "Роль пользователя (Админ, Читатель, Автор)"
          format: int
          example: 1
    UserFullInfo:
      required:
        - id
        - fio
        - login
        - registrationDate
        - password
        - role
      type: object
      properties:
        id:
          type: integer
          format: int
          example: 10
        fio:
          type: string
          example: Иванов Иван Иванович
        registrationDate:
          type: string
          example: 01.01.2001
        login:
          type: string
          example: login
        password:
          type: string
          example: password
        role:
          type: integer
          description: "Роль пользователя (Админ, Читатель, Автор)"
          format: int
          example: 1

    #####################################################################
    # Team schemas
    #####################################################################
...
  securitySchemes:
    JWTAuth:
      type: apiKey
      name: token
      in: header